Rename Label component to InputLabel
InputLabel will be a comoponent used in conjunction with the TextInput and Dropdown components only. The Checkbox and RadioButton components will use a different label component, CheckLabel. This is for formatting purposes.
