+++ src/Forms/Dropdown.tsx
@@ -8,7 +8,7 @@ import React, {
 import styled, { ThemeContext } from 'styled-components';
 import { fromTheme } from '../Theme';
 import ValidationErrorMessage from './ValidationErrorMessage';
-import Label, { POSITION } from './Label';
+import InputLabel, { POSITION } from './InputLabel';
 
 interface DropdownOptionProps {
   /** The label of the dropdown option */
@@ -69,7 +69,7 @@ const Dropdown: FunctionComponent<DropdownProps> = (props): ReactElement => {
   } = props;
   const theme = useContext(ThemeContext);
   return (
-    <Label
+    <InputLabel
       htmlFor={id}
       label={label}
       labelPosition={labelPosition}
@@ -104,7 +104,7 @@ const Dropdown: FunctionComponent<DropdownProps> = (props): ReactElement => {
           {errorMessage}
         </ValidationErrorMessage>
       )}
-    </Label>
+    </InputLabel>
   );
 };
